+++ src/gui/EscaperPanel.cpp
@@ -86,7 +86,7 @@ EscaperPanel::EscaperPanel(wxFrame* parent)
 				wxTheClipboard->SetData(new wxTextDataObject(str));
 				wxTheClipboard->Close();
 				MyFrame* frame = ((MyFrame*)(wxGetApp().GetTopWindow()));
-				frame->pending_msgs.push_back({ (uint8_t)StringEscaped });
+				frame->pending_msgs.push_back({ StringEscaped });
 			}
 		});
 }
